+++ e_metrobus/navigation/views.py
@@ -342,9 +342,7 @@ class ShareScoreView(TemplateView):
 
     def get_context_data(self, **kwargs):
         context = super(ShareScoreView, self).get_context_data(**kwargs)
-        context["points"] = get_object_or_404(
-            models.Score, hash=kwargs["hash"]
-        ).score
+        context["points"] = get_object_or_404(models.Score, hash=kwargs["hash"]).score
         return context
 
 
@@ -418,3 +416,28 @@ class FeedbackView(TemplateView):
         else:
             return self.render_to_response(self.get_context_data(feedback=feedback))
         return redirect("navigation:finished_quiz")
+
+
+class BugView(NavigationView):
+    template_name = "navigation/bug.html"
+    footer_links = {
+        "info": {"enabled": True},
+        "dashboard": {"enabled": True},
+        "leaf": {"enabled": True},
+        "results": {"enabled": True},
+    }
+
+    def get_context_data(self, **kwargs):
+        context = super(BugView, self).get_context_data(**kwargs)
+        context["bug"] = kwargs.get(
+            "bug", forms.BugForm(initial={"type": models.Bug.TECHNICAL}),
+        )
+        return context
+
+    def post(self, request, **kwargs):
+        bug = forms.BugForm(request.POST)
+        if bug.is_valid():
+            bug.save()
+        else:
+            return self.render_to_response(self.get_context_data(bug=bug))
+        return redirect("navigation:dashboard")