+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project tries to adhere to [Semantic Versioning](https://semver.org/spe
 ### Changed 
 - progress bar/cirle shows correct and wrong answers
 - added share link on navigation bar
+- used tabs in favor of accordion on questions-as-text page
  
 ## [0.10.0] - 2020-06-09
 ### Added