+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project tries to adhere to [Semantic Versioning](https://semver.org/spe
 - progress bar/cirle shows correct and wrong answers
 - added share link on navigation bar
 - used tabs in favor of accordion on questions-as-text page
+- footer icons and added labels
  
 ## [0.10.0] - 2020-06-09
 ### Added